+++ nav2_behavior_tree/test/plugins/action/test_bt_action_node.cpp
@@ -25,7 +25,7 @@
 
 #include "behaviortree_cpp/bt_factory.h"
 #include "nav2_behavior_tree/bt_action_node.hpp"
-#include "nav2_util/bt_loop_rate.hpp"
+#include "nav2_behavior_tree/utils/loop_rate.hpp"
 
 #include "test_msgs/action/fibonacci.hpp"
 
@@ -261,7 +261,7 @@ TEST_F(BTActionNodeTestFixture, test_server_timeout_success)
   BT::NodeStatus result = BT::NodeStatus::RUNNING;
 
   // BT loop execution rate
-  nav2_util::LoopRate loopRate(10ms, tree_.get());
+  nav2_behavior_tree::LoopRate loopRate(10ms, tree_.get());
 
   // main BT execution loop
   while (rclcpp::ok() && result == BT::NodeStatus::RUNNING) {
